# apps/api/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Local Market API",
    "description": "API publique de découverte des exploitations et de leurs offres.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/admin/farms/{farmId}": {
      "get": {
        "operationId": "getAdminFarm",
        "tags": ["admin"],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "farmId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "patch": {
        "operationId": "updateAdminFarm",
        "tags": ["admin"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 160,
                    "type": "string"
                  },
                  "description": {
                    "anyOf": [
                      {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "publicEmail": {
                    "anyOf": [
                      {
                        "format": "email",
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "publicPhone": {
                    "anyOf": [
                      {
                        "maxLength": 40,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                },
                "minProperties": 1
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "farmId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/farms/{farmId}/listings": {
      "get": {
        "operationId": "getAdminListings",
        "tags": ["admin"],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "farmId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "operationId": "createAdminListing",
        "tags": ["admin"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productCatalogId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "title": {
                    "minLength": 1,
                    "maxLength": 160,
                    "type": "string"
                  },
                  "description": {
                    "anyOf": [
                      {
                        "maxLength": 2000,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "variety": {
                    "anyOf": [
                      {
                        "maxLength": 120,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "unit": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": ["piece"]
                      },
                      {
                        "type": "string",
                        "enum": ["kilogram"]
                      },
                      {
                        "type": "string",
                        "enum": ["gram"]
                      },
                      {
                        "type": "string",
                        "enum": ["bunch"]
                      },
                      {
                        "type": "string",
                        "enum": ["box"]
                      },
                      {
                        "type": "string",
                        "enum": ["basket"]
                      }
                    ]
                  },
                  "unitQuantity": {
                    "pattern": "^\\d+(?:\\.\\d{1,3})?$",
                    "type": "string"
                  },
                  "priceCents": {
                    "minimum": 0,
                    "type": "integer"
                  },
                  "vatRate": {
                    "pattern": "^\\d+(?:\\.\\d{1,2})?$",
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "productCatalogId",
                  "title",
                  "unit",
                  "unitQuantity",
                  "priceCents",
                  "vatRate"
                ]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "farmId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/farms/{farmId}/listings/{listingId}": {
      "get": {
        "operationId": "getAdminListing",
        "tags": ["admin"],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "farmId",
            "required": true
          },
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "listingId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "patch": {
        "operationId": "updateAdminListing",
        "tags": ["admin"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "description": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "variety": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "priceCents": {
                    "minimum": 0,
                    "type": "integer"
                  },
                  "vatRate": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                },
                "minProperties": 1
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "farmId",
            "required": true
          },
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "listingId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/farms/{farmId}/inventory": {
      "get": {
        "operationId": "getAdminInventory",
        "tags": ["admin"],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "farmId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/farms/{farmId}/inventory/movements": {
      "post": {
        "operationId": "createStockMovement",
        "tags": ["admin"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "inventoryBatchId": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "type": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": ["stock_added"]
                      },
                      {
                        "type": "string",
                        "enum": ["stock_corrected"]
                      },
                      {
                        "type": "string",
                        "enum": ["stock_lost"]
                      },
                      {
                        "type": "string",
                        "enum": ["stock_refunded"]
                      }
                    ]
                  },
                  "quantity": {
                    "pattern": "^-?\\d+(?:\\.\\d{1,3})?$",
                    "type": "string"
                  },
                  "reason": {
                    "minLength": 1,
                    "maxLength": 500,
                    "type": "string"
                  }
                },
                "required": ["inventoryBatchId", "type", "quantity"]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "farmId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/farms/{farmId}/orders": {
      "get": {
        "operationId": "getAdminOrders",
        "tags": ["admin"],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "farmId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/farms/{farmId}/orders/{orderId}": {
      "patch": {
        "operationId": "updateAdminOrder",
        "tags": ["admin"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": ["preparing"]
                      },
                      {
                        "type": "string",
                        "enum": ["ready_for_pickup"]
                      },
                      {
                        "type": "string",
                        "enum": ["completed"]
                      },
                      {
                        "type": "string",
                        "enum": ["cancelled"]
                      },
                      {
                        "type": "string",
                        "enum": ["refunded"]
                      }
                    ]
                  }
                },
                "required": ["status"]
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "farmId",
            "required": true
          },
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "orderId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
...
}